refactor(bcf): split reader.ts/writer.ts under the module-size budget
reader.ts and writer.ts grew past their allowlisted budgets (1076/1058 and
939/900 lines) with PR #3970's three silent-drop fixes, blocking merge.

Extract BCFV viewpoint-content parsing (camera shapes, point/direction,
lines, clipping planes, bitmaps) out of reader.ts into
reader-viewpoint-content.ts, and the markup-element writers (header/file,
BIM snippet, document reference, line, clipping plane, bitmap) out of
writer.ts into writer-markup-elements.ts. Both extracted functions were
already unexported, so the package's public API surface is unchanged
(verified with check-api-surface.mjs). Pure refactor: 452/452 bcf tests
still pass, and all three of #3970's fixes were confirmed to still be
covered by mutating each back toward its pre-fix behaviour and observing
its test fail.

reader.ts: 1076 -> 804 lines (budget lowered to 804)
writer.ts: 939 -> 753 lines (budget lowered to 753)
