+++ scripts/module-size-allowlist.txt
@@ -185,9 +185,9 @@
    413 apps/viewer/vite.config.ts
    788 packages/bcf/src/ids-reporter.ts
    495 packages/bcf/src/overlay-renderer.ts
-  1058 packages/bcf/src/reader.ts
+   804 packages/bcf/src/reader.ts
    595 packages/bcf/src/viewpoint.ts
-   900 packages/bcf/src/writer.ts
+   753 packages/bcf/src/writer.ts
    605 packages/cache/src/glb.ts
    414 packages/clash/src/adapters/step.ts
    441 packages/clash/src/bcf-bridge.ts
